+++ src/components/autocomplete/username-selector.jsx
@@ -11,7 +11,7 @@ import { usePost } from '../post/post-comment-ctx';
 import { getMatchedUsers, showMoreComments } from '../../redux/action-creators';
 import style from './autocomplete.module.scss';
 import { HighlightText } from './highlight-text';
-import { withCustomEmojis } from '../custom-emoji';
+import { Twemoji } from '../twemoji';
 import {
   getAllGroups,
   getAllUsers,
@@ -100,7 +100,9 @@ function Item({ account, match, isCurrent, onClick, localLink }) {
           <HighlightText text={account.username} matches={match.matches} />
         </span>
         {account.username !== account.screenName && (
-          <span className={style.screenName}>{withCustomEmojis(account.screenName)}</span>
+          <span className={style.screenName}>
+            <Twemoji>{account.screenName}</Twemoji>
+          </span>
         )}
       </span>
       {localLink ? (