+++ src/components/linkify-elements.jsx
@@ -17,7 +17,6 @@ import {
   isShortLink,
   CODE_INLINE,
   CODE_BLOCK,
-  CUSTOM_EMOJI,
 } from '../utils/parse-text';
 import { INITIAL_CHECKBOX, isChecked } from '../utils/initial-checkbox';
 import UserName from './user-name';
@@ -26,7 +25,6 @@ import { Anchor, Link } from './linkify-links';
 import CodeBlock from './code-block';
 import { MediaLink } from './media-links/media-link';
 import { freefeedAttachmentId } from './media-links/helpers';
-import { CustomEmoji } from './custom-emoji';
 
 const { searchEngine } = CONFIG.search;
 const MAX_URL_LENGTH = 50;
@@ -158,9 +156,6 @@ export function tokenToElement(token, key, text, params) {
 
     case CODE_BLOCK:
       return <CodeBlock key={key} text={token.text} />;
-
-    case CUSTOM_EMOJI:
-      return <CustomEmoji key={key} emoji={token.text} />;
   }
   return token.text;
 }