+++ src/components/user-feed.jsx
@@ -6,7 +6,7 @@ import { Link } from '../services/nouter';
 import PaginatedView from './paginated-view';
 import Feed from './feed';
 import { SignInLink } from './sign-in-link';
-import { withCustomEmojis } from './custom-emoji';
+import { Twemoji } from './twemoji';
 
 class UserFeed extends Component {
   render() {
@@ -31,7 +31,7 @@ class UserFeed extends Component {
 
     const emptyFeedMessage = possiblyBlocked && (
       <p>
-        Perhaps <b>{withCustomEmojis(viewUser.screenName)}</b> has not written any posts yet
+        Perhaps <b><Twemoji>{viewUser.screenName}</Twemoji></b> has not written any posts yet
         {authenticated ? ' or they have blocked you' : ''}.
       </p>
     );
@@ -50,7 +50,7 @@ class UserFeed extends Component {
       return (
         <div className="box-body">
           <p>
-            You have blocked <b>{withCustomEmojis(viewUser.screenName)}</b>, so all of their posts
+            You have blocked <b><Twemoji>{viewUser.screenName}</Twemoji></b>, so all of their posts
             and comments are invisible to you.
           </p>
         </div>
@@ -62,12 +62,12 @@ class UserFeed extends Component {
       privacyMessage = (
         <div className="box-body">
           <p>
-            <b>{withCustomEmojis(viewUser.screenName)}</b> has a private feed.
+            <b><Twemoji>{viewUser.screenName}</Twemoji></b> has a private feed.
           </p>
           {!authenticated && (
             <p>
               <Link to="/signup">Sign up</Link> (or <SignInLink>sign in</SignInLink>) and request a
-              subscription to see posts from <b>{withCustomEmojis(viewUser.screenName)}</b>.
+              subscription to see posts from <b><Twemoji>{viewUser.screenName}</Twemoji></b>.
             </p>
           )}
         </div>
@@ -76,12 +76,12 @@ class UserFeed extends Component {
       privacyMessage = (
         <div className="box-body">
           <p>
-            <b>{withCustomEmojis(viewUser.screenName)}</b> has a protected feed. It is only visible
+            <b><Twemoji>{viewUser.screenName}</Twemoji></b> has a protected feed. It is only visible
             to {CONFIG.siteTitle} users.
           </p>
           <p>
             <Link to="/signup">Sign up</Link> or <SignInLink>sign in</SignInLink> to see posts from{' '}
-            <b>{withCustomEmojis(viewUser.screenName)}</b>.
+            <b><Twemoji>{viewUser.screenName}</Twemoji></b>.
           </p>
         </div>
       );
@@ -115,21 +115,21 @@ function UserGonePanel({ user }) {
     if (user.description) {
       return (
         <>
-          <b>{withCustomEmojis(user.screenName)}</b> has paused their account and left a message:{' '}
-          <em>{withCustomEmojis(user.description)}</em>
+          <b><Twemoji>{user.screenName}</Twemoji></b> has paused their account and left a message:{' '}
+          <em><Twemoji>{user.description}</Twemoji></em>
         </>
       );
     }
     return (
       <>
-        <b>{withCustomEmojis(user.screenName)}</b> has paused their account. They may return
+        <b><Twemoji>{user.screenName}</Twemoji></b> has paused their account. They may return
         someday.
       </>
     );
   }
   return (
     <>
-      <b>{withCustomEmojis(user.screenName)}</b> account has been deleted. This page still exists
+      <b><Twemoji>{user.screenName}</Twemoji></b> account has been deleted. This page still exists
       as a stub for the username, but this {user.type} is not in FreeFeed anymore.
     </>
   );