+++ src/utils/parse-text.js
@@ -15,11 +15,8 @@ import { linkHref } from 'social-text-tokenizer/prettifiers';
 import { makeToken, reTokenizer, wordAdjacentChars } from 'social-text-tokenizer/utils';
 import { withCharsAfter, withCharsBefore, withFilters } from 'social-text-tokenizer/filters';
 import { checkboxParser } from './initial-checkbox';
-import { CUSTOM_EMOJI, customEmojiTokenizer } from './emoji-overrides';
 import { SPOILER_END, SPOILER_START, spoilerTags, validateSpoilerTags } from './spoiler-tokens';
 import { isPostLink, isCommentLink } from './post-link-utils';
-
-export { CUSTOM_EMOJI } from './emoji-overrides';
 
 const {
   textFormatter: { tldList, foreignMentionServices },
@@ -124,8 +121,6 @@ const shortLinks = withFilters(
   withCharsAfter(wordAdjacentChars.withoutChars('/')),
 );
 
-const customEmojis = customEmojiTokenizer();
-
 export const lineBreaks = reTokenizer(/[^\S\n]*\n\s*/g, (offset, text) => {
   if (text.indexOf('\n') === text.lastIndexOf('\n')) {
     return makeToken(LINE_BREAK)(offset, text);
@@ -184,7 +179,6 @@ export const parseText = withTexts(
       lineBreaks,
       codeInline,
       codeBlocks,
-      customEmojis,
     ),
   ),
 );